feat: add custom highlighter option (#13)
At the moment, Rehype Pretty Code only allows the Shiki theme to
be configured. By adding a `getHighlighter` option, we can let users
configure any present or future Shiki options without having to
manually add support for them in Rehype Pretty Code.

For example, it is currently impossible to configure the `langs`
option. This change allows the user to pass in a highlighter
instance with the `langs` option already set.

The `getHighlighter` function is given the theme options prepared
by Rehype Pretty Code, allowing the custom highlighter to support
light and dark modes.