fix: don't cache highlighters
The problem that caching was aiming to solve is to avoid creating
highlighters for each code block in the given tree, but that doesn't
happen, without caching it's only being created once per tree, as the
new assertion proves.

The issue with this particular kind of caching is that two separate
compilers with different themes can only work properly if these themes
happen to be using different keys, otherwise they would overwrite each
other in the cache.

Probably a more common issue that this was causing was when changing
themes wouldn't be reflected in frontend frameworks like Remix. The
change would be visible only after restarting the process entirely.